Fix obsoletion warnings for JsonSerializerOptions.IgnoreNullValues
AzureProxyTool is serializing McpClientTool instances, which have a JsonSerializerOptions property, and so the source generator was emitting logic to serialize a JsonSerializerOptions. This was resulting in obsoletion warnings about obsolete properties on JSO (it was also resulting in unnecessary additional size for unnecessary source generated types and larger than necessary JSON).

It should instead be serializing the protocol-level Tool type.